Move generic sort functions from SortApartments to utils folder
